# Conflicts:
#	ap/train/trainer.py

+++ ap/train/data_manager.py
@@ -4,6 +4,7 @@
 
 import json
 import itertools
+import logging
 import os
 # import shutil
 # import tempfile
@@ -19,6 +20,7 @@ import numpy as np
 # import yaml
 
 # from ap.utils.general import batch_names, ensure_directory
+from ap.utils.general import recursively_unlink
 
 
 class NoTranslationException(Exception):
@@ -44,8 +46,11 @@ class ModelDataManager:
         self._config = experiment_config
 
         # self._data_dir = data_dir
-        self.train_grnti: typing.Dict[str, str] = self.get_rubric_of_train_docs()
-        self.train_dict: typing.Dict[str, str] = joblib.load(self._config["train_dict_path"])
+        self.train_grnti: typing.Dict[str, str] = self._get_rubric_of_train_docs()
+        # self.train_dict: typing.Dict[str, str] = joblib.load(self._config["train_dict_path"])
+        with open(self._config["train_vw_path"]) as file:
+            train_vw = file.readlines()
+        self.train_dict = {line.split()[0]: line for line in train_vw}
 
         path_experiment = Path(self._config["path_experiment"])
         path_experiment.mkdir(parents=True, exist_ok=True)
@@ -70,7 +75,7 @@ class ModelDataManager:
         # старые модальности - вытащить из модели
         # новые - из конфига
 
-        self._class_ids = experiment_config["LANGUAGES_TRAIN"]
+        self._class_ids = experiment_config["MODALITIES_TRAIN"]
         # self._class_ids_path = os.path.join(data_dir, "classes.yaml")
         # with open(self._class_ids_path, "r") as file:
         #     self._class_ids = yaml.safe_load(file)
@@ -121,7 +126,7 @@ class ModelDataManager:
     #     return artm.BatchVectorizer(data_path=[self._new_batches_dir, self._batches_dir],
     #                                 data_weight=[1, 1])
 
-    def get_rubric_of_train_docs(self):
+    def _get_rubric_of_train_docs(self):
         """
         Get dict where keys - document ids, value - number of GRNTI rubric of document.
 
@@ -315,16 +320,26 @@ class ModelDataManager:
         batches_list = list(self._path_to_batches.iterdir())
         if batches_list:
             for batch in batches_list:
-                batch.unlink()
+                if batch.is_file():
+                    batch.unlink()
+                else:
+                    recursively_unlink(batch)
         _ = artm.BatchVectorizer(
             data_path=str(self._path_balanced_train),
             data_format="vowpal_wabbit",
             target_folder=str(self._path_to_batches),
         )
-        batch_vectorizer = artm.BatchVectorizer(
-            data_path=[self._path_to_batches, self._path_batches_wiki],
-            data_weight=[1, 1]
-        )
+        if self._path_batches_wiki:
+            batch_vectorizer = artm.BatchVectorizer(
+                data_path=[self._path_to_batches, self._path_batches_wiki],
+                data_weight=[1, 1]
+            )
+            logging.info('Built batches with wiki')
+        else:
+            batch_vectorizer = artm.BatchVectorizer(
+                data_path=self._path_to_batches
+            )
+            logging.info('Built batches without wiki')
         return batch_vectorizer
 
     # def _merge_batches(self):
